feat(front): Save and restore path when logging in
